requirements.txt
faiss-cpu
numpy
pymupdf
nltk
langdetect
sentence-transformers
gradio
ollama
